+++ Assets/Scripts/EnemyBehaviours/RangedBehaviour.cs
@@ -44,7 +44,7 @@ public class RangedBehaviour : EnemyBehaviour
     {
         if(_isDamageReceived)
         {
-            _enemyInfo._health -= _currentPlayerDamage;
+            _enemyInfo._health -= (int)_currentPlayerDamage;
             _isDamageReceived = false;
             if(_enemyInfo._health <= 0)
             {
